# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: LiviPro B2B - API for Admin and Grossiste backoffices
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grossistes/{grossisteId}/stats:
    get:
      operationId: getGrossisteStats
      tags: [grossistes]
      summary: Get dashboard stats for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grossiste stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrossisteStats"
  /grossistes/{grossisteId}/chauffeurs:
    get:
      operationId: listChauffeurs
      tags: [chauffeurs]
      summary: List chauffeurs for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chauffeurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chauffeur"
    post:
      operationId: createChauffeur
      tags: [chauffeurs]
      summary: Create a chauffeur
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChauffeurInput"
      responses:
        "201":
          description: Created chauffeur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chauffeur"
  /grossistes/{grossisteId}/chauffeurs/{chauffeurId}:
    put:
      operationId: updateChauffeur
      tags: [chauffeurs]
      summary: Update a chauffeur
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: chauffeurId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChauffeurInput"
      responses:
        "200":
          description: Updated chauffeur
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chauffeur"
    delete:
      operationId: deleteChauffeur
      tags: [chauffeurs]
      summary: Delete a chauffeur
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: chauffeurId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /grossistes/{grossisteId}/boutiques:
    get:
      operationId: listBoutiques
      tags: [boutiques]
      summary: List boutiques for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of boutiques
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Boutique"
    post:
      operationId: createBoutique
      tags: [boutiques]
      summary: Create a boutique client
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoutiqueInput"
      responses:
        "201":
          description: Created boutique
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boutique"
  /grossistes/{grossisteId}/boutiques/{boutiqueId}:
    put:
      operationId: updateBoutique
      tags: [boutiques]
      summary: Update a boutique
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: boutiqueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBoutiqueInput"
      responses:
        "200":
          description: Updated boutique
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Boutique"
    delete:
      operationId: deleteBoutique
      tags: [boutiques]
      summary: Delete a boutique
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: boutiqueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /grossistes/{grossisteId}/produits:
    get:
      operationId: listProduits
      tags: [produits]
      summary: List products for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produit"
    post:
      operationId: createProduit
      tags: [produits]
      summary: Create a product
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduitInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
  /grossistes/{grossisteId}/produits/{produitId}:
    put:
      operationId: updateProduit
      tags: [produits]
      summary: Update a product
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: produitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProduitInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produit"
    delete:
      operationId: deleteProduit
      tags: [produits]
      summary: Delete a product
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: produitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /grossistes/{grossisteId}/tournees:
    get:
      operationId: listTournees
      tags: [tournees]
      summary: List delivery tours for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tournees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournee"
    post:
      operationId: createTournee
      tags: [tournees]
      summary: Create a delivery tour
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTourneeInput"
      responses:
        "201":
          description: Created tournee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournee"
  /grossistes/{grossisteId}/tournees/{tourneeId}:
    get:
      operationId: getTournee
      tags: [tournees]
      summary: Get tournee details
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: tourneeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourneeDetail"
    put:
      operationId: updateTourneeStatut
      tags: [tournees]
      summary: Update tournee status
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
        - name: tourneeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTourneeStatutInput"
      responses:
        "200":
          description: Updated tournee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournee"
  /grossistes/{grossisteId}/livraisons:
    get:
      operationId: listLivraisons
      tags: [livraisons]
      summary: List livraisons for a grossiste
      parameters:
        - name: grossisteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of livraisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Livraison"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    GrossisteStats:
      type: object
      properties:
        totalChauffeurs:
          type: integer
        totalBoutiques:
          type: integer
        totalTournees:
          type: integer
        tourneesEnCours:
          type: integer
        livraisonsAujourdHui:
          type: integer
        chiffreAffairesMensuel:
          type: number
        tauxLivraisonReussi:
          type: number
      required:
        - totalChauffeurs
        - totalBoutiques
        - totalTournees
        - tourneesEnCours
        - livraisonsAujourdHui
        - chiffreAffairesMensuel
        - tauxLivraisonReussi
...
    Chauffeur:
      type: object
      properties:
        id:
          type: integer
        grossisteId:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        permis:
          type: string
        statut:
          type: string
          enum: [disponible, en_tournee, inactif]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - grossisteId
        - nom
        - prenom
        - telephone
        - permis
        - statut
        - createdAt
    CreateChauffeurInput:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        permis:
          type: string
      required:
        - nom
        - prenom
        - telephone
        - permis
    UpdateChauffeurInput:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        permis:
          type: string
        statut:
          type: string
          enum: [disponible, en_tournee, inactif]
    Boutique:
      type: object
      properties:
        id:
          type: integer
        grossisteId:
          type: integer
        nom:
          type: string
        proprietaire:
          type: string
        adresse:
          type: string
        telephone:
          type: string
        limiteCredit:
          type: number
        soldeCredit:
          type: number
        statut:
          type: string
          enum: [actif, suspendu]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - grossisteId
        - nom
        - proprietaire
        - adresse
        - telephone
        - limiteCredit
        - soldeCredit
        - statut
        - createdAt
    CreateBoutiqueInput:
      type: object
      properties:
        nom:
          type: string
        proprietaire:
          type: string
        adresse:
          type: string
        telephone:
          type: string
        limiteCredit:
          type: number
      required:
        - nom
        - proprietaire
        - adresse
        - telephone
        - limiteCredit
    UpdateBoutiqueInput:
      type: object
      properties:
        nom:
          type: string
        proprietaire:
          type: string
        adresse:
          type: string
        telephone:
          type: string
        limiteCredit:
          type: number
        statut:
          type: string
          enum: [actif, suspendu]
    Produit:
      type: object
      properties:
        id:
          type: integer
        grossisteId:
          type: integer
        nom:
          type: string
        categorie:
          type: string
        prixUnitaire:
          type: number
        unite:
          type: string
        stockDisponible:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - grossisteId
        - nom
        - categorie
        - prixUnitaire
        - unite
        - stockDisponible
        - createdAt
...